Interceptor

+++ superherocrud/src/main/java/ar/com/vic/superherocrud/repositories/SuperHeroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ar.com.vic.superherocrud.model.SuperHero;
 
-public interface SuperHeroRepository extends JpaRepository<SuperHero, Integer> {
+public interface SuperHeroRepository extends JpaRepository<SuperHero, Long> {
 
 }
